requirements.txt
Django==4.2.23
cohere==4.47.0
requests==2.32.3
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
openai==1.95.1
opencv-python==4.12.0.88
openpyxl==3.1.2
torch==2.4.1
python-dotenv==1.0.1
django-allauth==0.61.1
pillow==10.3.0